✏️fix(EditarVentas): Se corrigo y mejor el editar ventas del modulo ventas, se cambio el diseñó y se cambio el diseñó de Detalle de ventas.

+++ CapaEntidades/DetalleVenta.cs
@@ -12,7 +12,7 @@ namespace CapaEntidades
         public int Id { get; set; }
         public int VentaId { get; set; }
         public int ProductoId { get; set; }
-        public decimal Cantidad { get; set; }
+        public int Cantidad { get; set; }
         public decimal SubTotal { get; set; }
         public bool Estado { get; set; }
